Add support for adding/removing other members on the map
We currently require users to have write access to the repository in
order add/remove them from the map on self-service.  In many case, this
is not ideal (e.g. contributors may have limited access to the
repository, former members cannot remove them) so add a field to allow
providing the login of the user to add or remove.

This still allow self-servicing for users with write access, but allow
these users to easily manage entries of other contributor without such
access.